Clean up warnings.
- Comparison of types with different signedness.
- Moved +[NSObject(PropertySupport) getPropertyType] up in file to drop private category and squelch "Category is implementing a method which will also be implemented by its primary class".
- Files missing a final newline.
- /* within a block comment.

+++ FayeObjc/Extensions/NSString+InflectionSupport.m
@@ -21,7 +21,7 @@
 	NSMutableString *underscored = [NSMutableString string];
 	
 	NSString *currChar;
-	for (int i = 0; i < [self length]; i++) {
+	for (NSUInteger i = 0; i < [self length]; i++) {
 		currChar = [NSString stringWithCharacters:buffer+i length:1];
 		if([[self capitals] characterIsMember:buffer[i]]) {
 			[underscored appendFormat:@"%@%@", delimiter, [currChar lowercaseString]];
@@ -55,7 +55,7 @@
 	
 	BOOL capitalizeNext = NO;
 	NSCharacterSet *delimiters = [self camelcaseDelimiters];
-	for (int i = 0; i < [self length]; i++) {
+	for (NSUInteger i = 0; i < [self length]; i++) {
 		NSString *currChar = [NSString stringWithCharacters:buffer+i length:1];
 		if([delimiters characterIsMember:buffer[i]]) {
 			capitalizeNext = YES;
